requirements.txt
nicegui
pandas
openpyxl
python-dotenv